Fix symmetry issue, accidentally dither iso (yay)
By moving around the "easeSquares()" step to before shading is
introduced, we effectively gain dithering on iso renders, and to a
lesser extent on ortho as well. Like all dithered pixel art, it looks
better without high zoom. The symmetry issue was that sometimes shading
would appear differently on the left side than the right; this seems to
be fixed by using the color indices to determine priority (any arbitrary
but deterministic way should work).