Generate, use alternate palette
There's a sample of 100 spaceships in an atlas saved in the repo (it's
about 328K, so nothing to worry about on size); these use this alternate
palette. The reason I prefer the alternate to the original is that it
smoothly ramps up perceptual brightness (saturation also changes
smoothly), whereas CW_PALETTE would mix desaturated and mixed-brightness
colors in the order from "darkest" to "lightest". The display test will
also use the new color scheme.