Complex code that can eliminate floating sections
This is used for the fish (press f in TestDisplay to make a fish, then
any directions will make a new one, such as numpad directions and also
hjklyubn vi-keys). I don't think fish or most lifeforms should have
disconnected parts, and this code can remove them when desirable while
keeping spaceships as-is. This code is in Tools3D, and mostly uses the
conventions SquidLib's GreasedRegion class does. Docs are forthcoming,
eventually, for Tools3D stuff.

Some rendering code was improved for fish-style eyes; now the glimmer in
them will show on all directions, not just isometric.

The sample atlas was moved out of the zip because zips are bad practice
to keep in repos, an opinion that I remind myself of now that it isn't 2
AM.