Serious tweaking for iso shading; much better
There shouldn't be the issue of unconnected shading lines like there was
before. This was related to the doubling in size of the color and depth
2D arrays, while the edges of voxels should have had lower depths than
the centers (instead, they were all the same). This discrepancy between
actual depth and used depth caused various shading issues, and seems to
be fixed now.