Experimenting with smaller iso renders (24x32)
They can't use the black outline that the other sizes do because the
outline needs to overlap with the voxel parts, and with 2x2 pixels used
for each voxel, that frequently covered a whole voxel with black outline
and hid its actual color. The solution actually looks rather good! You
can see the results in samples/ShadedSpaceSmaller.png , with 1000
isometric spaceships. It's a different style, more in the vein of some
early-to-mid-90s SEGA console games than the blockier black outlines
used in many NES games. The images themselves fit in a 24x32 space
(isometric, so some overlap is expected between tiles on the screen's
y-axis, and a floor sprite would be a diamond inscribed on a 24x24
square boundary).