Add option to render from below (at angle)
This uses the same approximate angle as from above (mirrored along the
plane parallel to the ground), but is dark from below instead of lit
from above, and the shading of the voxels, as well as the order they are
drawn, is different. You can activate "below mode" for ships or fish by
pressing 'b' in TestDisplay. Vi-keys aren't used any more; to make up
for this, both the number row keys and  the numpad keys can be used,
with the same meanings.