Big addition of side view renders
This was more involved than the below-45-degree renders, which still
needed some fixes and have been improved in this commit as well. In
TestDisplay, you can use 'b' to see from below, 's' to see from side,
'a' to see from above, and 'h' to cycle which one of these height angles
to use.