Drag and drop vox files to TestDisplay to render
This only works for MagicaVoxel .vox files that a) are the size this
expects (14x14x8 with an empty padding voxel on the
north/south/east/west sides) and b) use the right palette. This palette
is included in the repo now; if you're making .vox files to render with
this they should have the palette changed to WarpWriterPalette.png by
using the "Open" button below the palette in MagicaVoxel. This looks
like https://i.imgur.com/NjWDnFY.png , but has some weird issues at
other angles. The .vox file used to test is also in the repo,
Warrior_Male_W.vox . This switches to using LWJGL3 for drag and drop
support.