Fix homestar.vox palette, update palette saturation
All the non-grayscale colors in the default palette are now much more
saturated, which seems to work better with the retro appearance (most of
the NES and Amiga palettes are made of rather saturated colors, while
before we had mostly very desaturated colors). I'll keep trying to get a
good mix of saturated and desaturated colors, if possible. The
WarpWriterPalette.png file is updated, and with it are the palettes of
the other .vox models.