Render arbitrary-size .vox models; try Tree.vox
Tree.vox is an 80x80x80 voxel model  that can be dragged and dropped
onto TestDisplay's window to render it at its smallest size. It uses the
now-kinda-mandatory WarpWriterPalette.png for its palette; it's going to
be hard to have user-made palettes that fit the shading needs and the
current palette is mostly fine so far. The existing voxel models no
longer need empty padding on the sides, but the positioning isn't
perfect yet and some large models aren't fully shown (try pressing the
comma key to see a max-size block fail to fully render).