Fix remaining positioning issues
Tree.vox looks like this now: https://i.imgur.com/SrZL6Nw.png . The cube
models, accessible with `,` , now all are positioned in the same place
regardless of rotation, instead of being partly off-screen before. You
can also check this on the spaceships. The fish seem to have a glitch in
eye placement where their eyes are always touching and forwards-facing,
though it's OK for now. It could be because they're thinner now, so they
don't have room to place eyes to the side.