Color priority change helps small details in iso
Colors that are only rendered a few times in an image now are considered
to have higher priority when considering if they can be safely
overwritten. This prevents important details like eyes from being erased
simply for having a small size. Before, this was determined completely
by the order in the palette, but that isn't a very reasonable approach
when only one palette is expected to be used for now.