Fix shading on "priority" colors, side renders
Colors that are in the upper half of the palette have priority for
shading (they won't be overdrawn easily by non-priority nearby colors),
but clampDown(), which reduces brightness of a pixel to a minimum of the
darkest color in its hue/saturation block, wasn't working correctly with
the priority colors. It works now. The side-view ortho renders also were
miscalculating depth on some directions, leading to a very flat
appearance; this is also fixed.