Slight tweaks to palette, again
Before, the saturation was either correctly high or generally too low
for most things. Now the low-saturation parts of the palette have been
adjusted slightly. For example, low-saturation red is closer to brown
(but a rich wood-like brown instead of a grayish-brown or the brown
that's lighter, less saturated, and usable for some skin tones). Yellow
has less options that are unpleasantly grayish and dark, and more that
are cream-like (suitable for lights, maybe). The .vox models have been
updated to use the new palette and it especially improves hair. A sample
is in FarOutPalette , because it's after PaletteFromOuterSpace and today
is a far-out day.