Larger ships work, except for position
Like with the large tree model, the ortho renders aren't  positioned in
the same place as iso renders. TestDisplay seems to remake the canvaas
to fit smaller models correctly when a direction is specified, but not
automatically when a new model is made. TestDisplay currently defaults
to larger ships, but a switch should be added in soon to go back to
smaller ships. A sample of one ship is in samples/BigShipPreview/Uss/ .