Adjust larger ships so they are all one solid
This seems to have fixed a rather significant bug in
Tools3D.largestPart(), which is important here. The ships also have much
less empty space internally, but anything unconnected to the largest
part of the model (almost always the cockpit) is erased, and that
provides nice shape differences. It looks like
https://i.imgur.com/9bY8Uqx.gifv now; note the cockpit shading when it
faces the camera on some ships and how a grid of shading appears that
matches the slope.