Simple test to see how we can process .vox files
This test was written to produce sample data for @psilord to test on,
and since he's testing a Common Lisp library, the lambda-in-parentheses
logo-like model is a good thing to try to read in. This writes that logo
to a .lisp file, really just bare s-expressions containing
space-delimited 0 and 1 in nested parentheses. I have a feeling I'll be
modifying this test in the future to write other formats as they come
up.