Fix an issue with point hashing
Noise.PointHash in the latest SquidLib has some issues, which a different point hash here resolves. This also uses much better names for ships in TestOutput, like "FinalSultana" instead of "Eflupoull". There's a new sample that shows the big ships with the current point hashing (since it affects the shape and pattern) and naming.