Fix incorrect side-view positions
Tiny mode is commented out for now in TestDisplay because I don't think it has worked in a while. Large mode is enabled by default, you can disable it with 'l' or 'L'. Now the side view is correctly-positioned for ortho side views relative to the other ortho views. This also tweaks the point hash used to determine shape and colors for ships and fish, and allows making full-size debug blocks in large mode (use ',' to make a block, maybe with 'L' before that to toggle large mode to its other setting).