ByteFill threshold should be an int. Using Java name conventions.
ByteFill.transparent() was comparing the numerical value of a byte to the result of (getPixel(x, y)  & 255); a Java byte has, for bad reasons, a -128 to 127 range, and the `& 255` has a 0 to 255 range, so there would be no way to request a threshold that was above 127. I want to stick with the Java convention of type names being capitalized and method/variable/field names starting with a lower-case letter, so the constructor-like methods called Fill2D() and such I think should be fill2D().