Fix ship color issue, make easing of square shapes optional
For ships, colors could go past the darkest shade and wrap around to another hue at the lightest shade when it should have been the darkest. This is fixed. You can set ModelRenderer.easing to false or pass easing=false to the ModelRenderer constructor if you want to disable easing, which affects pixels based on their neighbors and the frequency at which both the colors of that pixel and the colors of its neighbors appear. Easing tends to improve how models look at larger sizes, but it can make debugging harder.