Keep tweaking point hashes
Since they probably will be used mostly for non-negative positions, it should be fine to use a function that behaves better when all components of a position are non-negative. It may be just fine for negative positions, it just may have duplicates more often, and while that matters for a hash table, we aren't using this in one of those. The function is a lot shorter and seems to yield good results on fish, blobs, and ships (especially ships, where it doesn't seem to produce the "cockpit only" glitchy ships at all).