Important bug fixes to ColorFetch
The colorFetch() method was being called before colors was initialized (transparent called colorFetch() before colors was assigned a value), but colors was accessed in colorFetch(). This might not be an issue sometimes, but static initialization order is a notoriously bad part of Java, and we don't know what would happen seemingly at random when the order is different from normal. Also, byte is signed in Java for more bad reasons, so accessing an array with a byte needs a bitwise AND mask to make it properly positive and within the right range.