Improved outlines, I think, with smoothing on and off
When smoothing is off you usually want to tell where edges are, as a debugging aid, and even when smoothing is on it can help to see more edges to tell which way a slope is heading and such. I managed to get the edges to show in the ortho modes so they don't get a dashed line sometimes, and still smooth well.

This also fixes a dumb mistake in point hashing where the signed right shift operator was used instead of the unsigned one.