Adjust order, color selection in Rinsed palette
The sorted palette is replaced with the regular RINSED palette, which now uses the correct order (sorted, but with an exception so practically-grayscale colors are where existing code expects grayscale colors to be). Models are adjusted to use this different palette. If you make a .vox model in MagicaVoxel, use Rinsed.png (in the top level here) as the palette.