Possible simplification based on how Roll should work
Since the axis that points in the direction of the viewer is x+ or x- (not sure which at this point), and roll always rotates around this axis, then we can rotate the face first without as complicated of a switch statement, and then do one quick switch at the end to rotate based on roll, which should be the same type of rotation (around x) every time. This still has size problems, and I'll update the voxel reference cube so it is easy to tell when an edge isn't being shown and the camera is inside the box (it is hollow).