Add VoxelModel to experimental warp package
This removes SparseVoxelSet and VoxModel because they were almost always going to be slower and use more memory than ArrayModel, plus they were totally unused.

The new VoxelModel class is a bit of an experiment in using SimpleDraw's style of rendering with a different approach to storing the voxel data, hopefully allowing rotations to actually work at some point. It may be possible to bridge the Fetch-based code with this to produce VoxelModels from a Fetch, but after debugging some of the Fetch-related code, I really don't want it running every frame.