Show FPS counter; surprisingly, Simple and Warp are the same
It is probably because we render relatively few voxels and scale the image up, so the actual calculations are being overshadowed by the overhead of just drawing one frame.

There's also some minor optimizations to VoxelModel; I really don't think VoxelModel should allow overriding the voxels() method, since it's pretty hard to modify correctly, so WarpDraw uses the voxels field, which seems a tiny bit faster since it's accessed in a tight loop.