Minor changes to noise-based dithering
This matches the changes in SquidLib yesterday, which seem to improve some areas when palette-reducing images of paintings. There's still way too much gritty, chaotic scattering of color (particularly reds and greens), so ditherSolid() seems to be the way to go for now, and it looks pretty nice in TestDisplay when you drop a .hex palette on the window.