Ramp-based AuroraTwilight works, but needs models that use it
The system for Aurora's colors in the original palette doesn't follow much of a pattern, so this lays out colors into ramps of 4 from light to dark. If a model isn't made using this slight variant on the Aurora palette, the colors will mean different things, and that's no good (it will still shade them correctly though, which isn't the case in Rinsed). This doesn't have any special placeholder colors, so it may be unusable in the current code anyway. The palette could be tweaked to make room for those placeholders and other special entries.