Mostly done with WarpDraw by adding iso view
It doesn't use triangles, so its perspective is slightly different from the version in VoxelDraw (the same is true for the above ortho rendering in WarpDraw vs. VoxelDraw), but it doesn't need "mixels" to stretch its pixels to match the right size. The smoothing/easing is probably going to need tweaking still because it mashes up details like the text on dumbcube.vox.