Add Colorizer, change some names (Twilight variables to Dimmer)
Colorizer implements IColorizer and extends Dimmer (large parts of which it gets essentially for free by having a PaletteReducer in it). IDimmer.twilight() is renamed to IDimmer.medium() because it kinda makes some sense with bright(), dim(), and dark(). RAMPS and similar constants now say what palette they use, so that is now AURORA_RAMPS. PaletteReducer now exposes reduceIndex(), which was possible to do manually before but is now easier.