Allow some palettes to change their shading rules with special indices
This is pretty much handled already by the existing outline/ramp mechanics in Colorizer, and only needs a method to allow marking colors as using the alternate shading rules. The plan is to use this to implement fire, smoke, dust, and glowing lights, none of which should have outlines or have the appearance of shading.