Adjust int hashes in ModelMaker so there are more possible outputs
Before, about 1/3 of all ints wouldn't appear in the output, though that's just an estimate and it could have been not that bad. Now, the adjusted coordinates are used again in the last step, which should help randomize the output much more. TestDisplay uses the int hashes as part of the ships it generates, so you can compare ships there to ones that also use FastNoise to smooth the design in WarpTest.