Reduce unwanted patterns in ModelMaker's point hashes
The latest 2D version was checked out in SquidLib's big HashVisualizer, and while the old last step of hashAll() wasn't enough to really randomize an int (lots of repetitive "wave" patterns showed up in the visualizer), the new step is. Thanks to Pelle Evensen for showing that double-shifts with xor are useful and bijective, even though these hashes are not bijections.