Add BurstTest to see if WarpDraw is the bottleneck... it isn't.
This also found a weird bug when drawing against the back edges in VoxelDraw using the 45 degree side view. Drawng a 3x3 grid of slopes, the front center slope should fully obscure the slopes behind it; it obscures the center slope but not the slope behind it: https://vgy.me/IkFifs.png . There's extra voxel halves and sometimes extra voxels drawn next to voxels at the edge, on both the left and right sides. Any ideas, @benmclean ? I'm pretty lost on this particular bug. Could it be clamping the voxel position in any relevant part of the code, instead of treating an outside-bounds voxel as 0?