Changes to arbitraryColorizer dark outlines so they aren't... weird and bad
The outlines were calculated incorrectly in the last commit, using values from the middle-lightness color to compare the darkest outline color to, instead of the middle-darker color that it should have based the outline on. This sometimes made outlines brightly oversaturated instead of dark and... outline-like. An extra step of desaturating the outlines is done now, too.