Add IntByteOrderedMap and IntSort
IntByteOrderedMap is based on SquidLib IntDoubleOrderedMap, and is meant to be used for voxel models (compressing x, y, and z into one int key using HashMap3D.fuse() ) that store only voxels that could need to be drawn, can be sorted when they change, and when sorted correctly back-to-front could be just rendered by iterating through the IntByteOrderedMap (likely using keyAt() and getAt() ). This would skip all empty voxels entirely, a major source of overhead currently on larger models.