Update color difference metric for PaletteReducer; improves color choices
A relatively brute-force brightness measure is incorporated into PaletteReducer.difference() and used to differentiate colors that have different-enough brightness to overshadow similar hue and saturation (since the eye is very sensitive to any brightness changes, but less so to subtle hue changes). This changes the preload codes in FlesurrectReducer and RinsedColorizer.