Add some larger models, one with effects, for testing
In palettes/, there's also a specialized palette, FlesurrectBonusMagicaVoxel.png , meant for drawing models that have effects in use for Flesurrect or FlesurrectBonus; it still uses the same colors as FlesurrectBonus, but can be used for Flesurrect as well. The first 64 colors (at the bottom in MagicaVoxel's palette) are normal colors, above that are solid unshaded colors, then wavering colors (for liquids and such), then shining/blinking colors at the top.