First draft of TurnQuaternion with methods for converting brads
It's easier and more precise internally to deal with small floats than large ones, so using turns from 0 to 1 internally should be helpful. TurnQuaternion.bradToTurn() should be the main way of getting the turns if we have brads. The inverse, turnToBrad(), is a little more finicky about its input, and maybe shouldn't be relied on heavily (it handles non-negative turns fine, though).