Add/update angle-related methods in TurnQuaternion
It now includes ways to lerp between angles measured in turns or brads. The terms for yaw, pitch, and roll now refer to the different axes used by WarpWriter relative to libGDX; they still represent the same physical rotation directions, just the labels are different on the axes. I use a mnemonic where roll is the direction a ship should make when you hear "FOX, DO A BARREL ROLL", pitch is the arcing lunge motion of a baseball pitcher (or softball, in the reverse direction), and yaw is the sound you make when your car is sliding uncontrollably on slick ice.