Add part of Ripple palette; not quite as good as Judge for most usage
Ripple uses colors selected by repeatedly cycling through the hue in HCL colorspace, with chroma and lightness selected by variations on sine waves that change (usually) with each cycle to a similar hue. Chroma.js was used externally to generate the color list, and OverkillPaletteGenerator was used to trim the list of colors down to 63. Most of the time, Judge does better in a similar domain, though.