VoxelSpotRenderer shows each voxel as one pixel (very small)
This needed some changes to VoxelColor, but they could be moved to the Spot renderer itself. It's weird because it isn't technically rendering rectangles or triangles, but instead uses GL_POINTS, so the rectangle and triangle renderers don't make sense, and scaling doesn't really make sense either. An advantage here is that this naturally makes sense for the colors to change per spot/pixel in more flexible ways, and a shader could later force a palette if we want.