Hope this helps... Ryan.
Labrador256 is a decent palette anyway, I just don't have the computing power to give it a larger color set to work with. This used a 9x9x9 grid of colors and minimized it using CIELAB metrics to determine what was closest. See Labrador256.png in palettes/ for an easy result.