Poked with stick, VoxelImmediateRenderer has issues.
I don't know exactly what's wrong, but I think this is the Windows 10 issue Ben experienced. I still have no idea what the cause is. Only the demos that use VoxelImmediateRenderer are affected, as far as I can see. I updated libGDX to the current 1.9.13 and SquidLib to 3.0.3, which is slightly behind libGDX but is the most recent stable version.